Reports & deletes
Moved report models to own folder & made own controller. Finished all reports except daily shipments (date restricts not working yet)

All other deletes (agent, sender, vehicle type, package type, access role) - partly working

Fixed Exception ex not returning anything

+++ CORE_WebAPI/Models/Reports/7IncomeReport.cs
@@ -9,11 +9,11 @@ namespace CORE_WebAPI.Models.Reports
     {
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public double TotalIncome { get; set; }
-        public double ShipmentIncome { get; set; }
-        public double PenaltyIncome { get; set; }
-        public double AgentPayments { get; set; }
-        public double ProfitLoss { get; set; }      
+        public decimal TotalIncome { get; set; }
+        public decimal ShipmentIncome { get; set; }
+        public decimal PenaltyIncome { get; set; }
+        public decimal AgentPayments { get; set; }
+        public decimal ProfitLoss { get; set; }
         public bool isProfit { get; set; }
         public string EmpFullName { get; set; }
     }
